Separated encryption types
