--- LocalKeyClientEncryption/LocalKeyClientEncryption.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Azure.Identity" Version="1.1.1" />
-    <PackageReference Include="Azure.Security.KeyVault.Keys" Version="4.0.3" />
-    <PackageReference Include="Azure.Storage.Blobs" Version="12.5.0-preview.1" />
-    <PackageReference Include="System.Configuration.Abstractions" Version="2.0.2.45" />
-  </ItemGroup>
-
-</Project>
