+++ keyVaultClientSideToCustomerManagedServerSide/keyVaultClientSideToCustomerManagedServerSide.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30204.135
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "keyVaultClientSideToCustomerManagedServerSide", "keyVaultClientSideToCustomerManagedServerSide.csproj", "{D1DE608C-88F2-4722-885D-5E7755127812}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{D1DE608C-88F2-4722-885D-5E7755127812}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D1DE608C-88F2-4722-885D-5E7755127812}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D1DE608C-88F2-4722-885D-5E7755127812}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D1DE608C-88F2-4722-885D-5E7755127812}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {A7F74BB5-8A7C-45E3-8225-834BBBE16AA4}
+	EndGlobalSection
+EndGlobal
