+++ keyVaultClientSideToCustomerManagedServerSide/samplesrc/Constants.cs
@@ -1,0 +1,11 @@
+﻿namespace keyVaultClientSideToCustomerManagedServerSide
+{
+    public class Constants
+    {
+        //Edit the Following as Needed
+        //Program Creates the Following using the Provided Names
+        public const string containerName = "example";
+        public const string fileName = "example.txt";
+        public const string encryptionScopeName = "myencryption";
+    }
+}
