+++ keyVaultClientSideToCustomerProvidedServerSide/keyVaultClientSideToCustomerProvidedServerSide.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30204.135
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "keyVaultClientSideToCustomerProvidedServerSide", "keyVaultClientSideToCustomerProvidedServerSide.csproj", "{CD5CA976-4A19-4D34-B48D-B64300859569}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{CD5CA976-4A19-4D34-B48D-B64300859569}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CD5CA976-4A19-4D34-B48D-B64300859569}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CD5CA976-4A19-4D34-B48D-B64300859569}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CD5CA976-4A19-4D34-B48D-B64300859569}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {2974E1C8-161A-400E-9C79-245B6C4FDD8B}
+	EndGlobalSection
+EndGlobal
