+++ keyVaultClientSideToMicrosoftManagedServerSide/keyVaultClientSideToMicrosoftManagedServerSide.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30204.135
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "keyVaultClientSideToMicrosoftManagedServerSide", "keyVaultClientSideToMicrosoftManagedServerSide.csproj", "{E0E34581-2FC4-411D-9C82-940E35A00695}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{E0E34581-2FC4-411D-9C82-940E35A00695}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E0E34581-2FC4-411D-9C82-940E35A00695}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E0E34581-2FC4-411D-9C82-940E35A00695}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E0E34581-2FC4-411D-9C82-940E35A00695}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {9280817D-FC41-4813-85C3-048B4E6A6BA8}
+	EndGlobalSection
+EndGlobal
