+++ localKeyClientSideToCustomerProvidedServerSide/localKeyClientSideToCustomerProvidedServerSide.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30204.135
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "localKeyClientSideToCustomerProvidedServerSide", "localKeyClientSideToCustomerProvidedServerSide.csproj", "{93BAA0DA-B72E-46D6-94BC-2A6D04EEC3B1}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{93BAA0DA-B72E-46D6-94BC-2A6D04EEC3B1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{93BAA0DA-B72E-46D6-94BC-2A6D04EEC3B1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{93BAA0DA-B72E-46D6-94BC-2A6D04EEC3B1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{93BAA0DA-B72E-46D6-94BC-2A6D04EEC3B1}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {D86A4207-BE96-4C6B-968C-DA44B1D1466F}
+	EndGlobalSection
+EndGlobal
