+++ localKeyClientSideToCustomerProvidedServerSide/samplesrc/Constants.cs
@@ -1,0 +1,11 @@
+﻿namespace localKeyClientSideToCustomerProvidedServerSide
+{
+    class Constants
+    {
+        //Edit the Following as Needed
+        //Program Creates the Following using the Provided Names
+        public const string containerName = "example";
+        public const string fileName = "example.txt";
+        public const string customerProvidedKey = "dfD3Jb#6htqfpoj@gGpomDAv21035%21";   //Key used for Customer Provided Key Server Side Encryption
+    }
+}
